Feature: Add display of PDL objects in 'x' commands
PDL objects (ndarrays) are scalar references, and PDL offers a nice,
configurable stringification for their objects.  This stringification
is available to the debugger's 'p' command.

With this patch, the stringification is also used when a PDL object
is shown with the 'x' command.  This is useful for PDL objects which
are part of a hash or array, or attributes of objects.